changed mapping for external items to float

+++ app/src/main/java/com/daemo/myfirsttrip/fragments/CostsListFragment.java
@@ -48,7 +48,7 @@ public class CostsListFragment extends ListFragment {
     }
 
     @Override
-    protected void setItemRelatedIds(Map<String, Integer> selectedIds) {
+    protected void setItemRelatedIds(Map<String, Float> selectedIds) {
         if (person != null)
             person.setCostsIds(selectedIds);
         else if (trip != null)
